+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * BREAKING: Removed archive support.
 * BREAKING: `DirectoryDescriptor.load` only supports a `String` path instead of
   also accepting relative `Uri` objects.
+* BREAKING: `DirectoryDescriptor.load` no longer has an optional `parents`
+  parameter - this was intended for internal use only.
 
 ## 1.2.0